--- server/client/src/Components/Player.js
@@ -1,11 +1,0 @@
-import React from 'react'
-
-function Player() {
-    return (
-        <div>
-           <h2>Player</h2> 
-        </div>
-    )
-}
-
-export default Player
