+++ server/client/src/Components/styles.css
@@ -15,7 +15,7 @@ h3 {
 .home-ctr {
     height: 1000px;
     width: 100%;
-    background: url('../../images/home-background.png');
+    background: url('../images/home-background.png');
     display: flex;
     flex-direction: column;
     align-items: center;
