+++ server/client/src/pages/HomePage/index.js
@@ -1,44 +1,14 @@
 import React from "react";
-import CreatePlaylist from "../../Components/CreatePlaylist";
-import Search from "../../Components/Search";
-import Welcome from "../../Components/Welcome";
-import PopularPlaylist from "../../Components/PopularPlaylist";
-import logo from '../../images/logo.png'
-import "./styles.css"
-import SearchResults from "../../Components/SearchResults";
-import PlaylistTitle from "../../Components/PlaylistTitle";
-import SongDetails from "../../Components/SongDetails";
-import YourPlaylist from "../../Components/YourPlaylist";
-import Player from "../../Components/Player";
+import VibezPlayer from "../../Components/VibezPlayer";
+import SpotifyWebApi from 'spotify-web-api-js';
+
+
+const spotify  = new SpotifyWebApi();
 
 const HomePage = () => {
 
 return (
-    <div className="home-ctr"> 
-        <div className="header">
-           <img className="logo" src={logo} alt="website logo"/>  
-        </div>
-        <div className="body">
-            <div className="sect-1">
-                <Welcome />
-                <CreatePlaylist />
-                <Search />
-                <SearchResults />
-                <PlaylistTitle />
-                <SongDetails />
-                <SongDetails />
-                <SongDetails />
-                <SongDetails />
-            </div>
-            <div className="sect-2">
-                <PopularPlaylist />
-                <YourPlaylist />
-            </div>
-            </div>
-        <div className="play-footer">
-                <Player />
-            </div>
-    </div>
+    <VibezPlayer />
     )
 }
